# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  error?: string;
}

export interface SuccessResponse {
  message: string;
  success: boolean;
}

export interface SignupRequest {
  name: string;
  email: string;
  phone: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  kycCompleted: boolean;
  creditScore?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type LoanProductType =
  (typeof LoanProductType)[keyof typeof LoanProductType];

export const LoanProductType = {
  home: "home",
  personal: "personal",
  business: "business",
  education: "education",
  vehicle: "vehicle",
  gold: "gold",
} as const;

export interface LoanProduct {
  id: number;
  name: string;
  type: LoanProductType;
  description: string;
  minAmount: number;
  maxAmount: number;
  minTenureMonths: number;
  maxTenureMonths: number;
  baseInterestRate: number;
  processingFeePercent: number;
  eligibilityCriteria: string[];
  documentsRequired: string[];
  isActive: boolean;
  createdAt: string;
}

export interface EmiCalculationRequest {
  loanAmount: number;
  tenureMonths: number;
  interestRate: number;
}

export interface EmiCalculationResult {
  monthlyEmi: number;
  totalAmount: number;
  totalInterest: number;
  processingFee: number;
  loanAmount: number;
  tenureMonths: number;
  interestRate: number;
}

export type LoanApplicationStatus =
  (typeof LoanApplicationStatus)[keyof typeof LoanApplicationStatus];

export const LoanApplicationStatus = {
  pending: "pending",
  kyc_review: "kyc_review",
  document_review: "document_review",
  underwriting: "underwriting",
  sanctioned: "sanctioned",
  rejected: "rejected",
  locked: "locked",
  disbursed: "disbursed",
} as const;

export type LoanApplicationAgentStage =
  | (typeof LoanApplicationAgentStage)[keyof typeof LoanApplicationAgentStage]
  | null;

export const LoanApplicationAgentStage = {
  sales: "sales",
  verification: "verification",
  underwriting: "underwriting",
  sanction: "sanction",
  locking: "locking",
} as const;

export interface LoanApplication {
  id: number;
  userId: number;
  loanProductId: number;
  loanProductName: string;
  loanAmount: number;
  tenureMonths: number;
  monthlyEmi: number;
  status: LoanApplicationStatus;
  agentStage?: LoanApplicationAgentStage;
  lockingId?: string | null;
  creditScore?: number | null;
  remarks?: string | null;
  applicantName: string;
  applicantEmail: string;
  applicantPhone: string;
  annualIncome?: number | null;
  employmentType?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateApplicationRequest {
  loanProductId: number;
  loanAmount: number;
  tenureMonths: number;
  monthlyEmi: number;
  annualIncome?: number;
  employmentType?: string;
}

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  pending: "pending",
  kyc_review: "kyc_review",
  document_review: "document_review",
  underwriting: "underwriting",
  sanctioned: "sanctioned",
  rejected: "rejected",
  locked: "locked",
  disbursed: "disbursed",
} as const;

export type UpdateStatusRequestAgentStage =
  | (typeof UpdateStatusRequestAgentStage)[keyof typeof UpdateStatusRequestAgentStage]
  | null;

export const UpdateStatusRequestAgentStage = {
  sales: "sales",
  verification: "verification",
  underwriting: "underwriting",
  sanction: "sanction",
  locking: "locking",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
  agentStage?: UpdateStatusRequestAgentStage;
  remarks?: string | null;
  lockingId?: string | null;
}

export interface ChatMessageRequest {
  message: string;
  sessionId: string;
  applicationId?: number | null;
}

export type ChatMessageResponseAgentType =
  (typeof ChatMessageResponseAgentType)[keyof typeof ChatMessageResponseAgentType];

export const ChatMessageResponseAgentType = {
  master: "master",
  sales: "sales",
  verification: "verification",
  underwriting: "underwriting",
  sanction: "sanction",
  locking: "locking",
} as const;

export interface ChatMessageResponse {
  response: string;
  agentType: ChatMessageResponseAgentType;
  nextStep?: string | null;
  options?: string[] | null;
  applicationId?: number | null;
  requiresAction?: string | null;
  sessionId: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  agentType?: string | null;
  sessionId: string;
  createdAt: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface DashboardSummary {
  totalApplications: number;
  pendingApplications: number;
  approvedApplications: number;
  rejectedApplications: number;
  totalLoanAmount: number;
  totalUsers: number;
  recentApplications: LoanApplication[];
  applicationsByStatus: StatusCount[];
  applicationsByType: TypeCount[];
}

export type CreateLoanProductRequestType =
  (typeof CreateLoanProductRequestType)[keyof typeof CreateLoanProductRequestType];

export const CreateLoanProductRequestType = {
  home: "home",
  personal: "personal",
  business: "business",
  education: "education",
  vehicle: "vehicle",
  gold: "gold",
} as const;

export interface CreateLoanProductRequest {
  name: string;
  type: CreateLoanProductRequestType;
  description: string;
  minAmount: number;
  maxAmount: number;
  minTenureMonths: number;
  maxTenureMonths: number;
  baseInterestRate: number;
  processingFeePercent: number;
  eligibilityCriteria: string[];
  documentsRequired: string[];
  isActive: boolean;
}