# lib/api-zod/src/generated/types/chatMessageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageRequest {
  message: string;
  sessionId: string;
  applicationId?: number | null;
}