# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  kycCompleted: boolean;
  creditScore?: number | null;
  createdAt: Date;
}